2017 Complete
The last images from 2017! Note: MCG-05-23-016 would not self-cal for me, so I kept the 1" and 6" images that Krista Smith made. If needed, I can try to download the calibrated measurement set from NRAO, and make a 3" image. I believe the reason I could not self-cal this source is due to version of CASA that created the measurement set.